+++ nas-20170626/alibabacloud_nas20170626/models/_describe_file_systems_response_body.py
@@ -768,8 +768,10 @@ class DescribeFileSystemsResponseBodyFileSystemsFileSystemPackagesPackage(DaraMo
 class DescribeFileSystemsResponseBodyFileSystemsFileSystemOptions(DaraModel):
     def __init__(
         self,
+        enable_abe: bool = None,
         enable_oplock: bool = None,
     ):
+        self.enable_abe = enable_abe
         # Specifies whether to enable the oplock feature. Valid values:
         # 
         # *   true: enables the feature.
@@ -786,6 +788,9 @@ class DescribeFileSystemsResponseBodyFileSystemsFileSystemOptions(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.enable_abe is not None:
+            result['EnableABE'] = self.enable_abe
+
         if self.enable_oplock is not None:
             result['EnableOplock'] = self.enable_oplock
 
@@ -793,6 +798,9 @@ class DescribeFileSystemsResponseBodyFileSystemsFileSystemOptions(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('EnableABE') is not None:
+            self.enable_abe = m.get('EnableABE')
+
         if m.get('EnableOplock') is not None:
             self.enable_oplock = m.get('EnableOplock')